+++ apps/mobile/ios/PokerTimer/AppDelegate.swift
@@ -20,6 +20,8 @@ public class AppDelegate: ExpoAppDelegate {
     reactNativeDelegate = delegate
     reactNativeFactory = factory
 
+    LiveActivityActionListener.shared.start()
+
 #if os(iOS) || os(tvOS)
     window = UIWindow(frame: UIScreen.main.bounds)
     factory.startReactNative(